refactor: More classes that needed updated import statements, update MockFileFinder to reflect new FileFinder API.
