+++ common/src/main/java/com/synopsys/integration/common/util/finder/FileFinder.java
@@ -1,5 +1,5 @@
 /*
- * detectable
+ * common
  *
  * Copyright (c) 2021 Synopsys, Inc.
  *
