+++ common/src/main/java/com/synopsys/integration/common/util/finder/WildcardFileFinder.java
@@ -1,3 +1,10 @@
+/*
+ * common
+ *
+ * Copyright (c) 2021 Synopsys, Inc.
+ *
+ * Use subject to the terms and conditions of the Synopsys End User Software License and Maintenance Agreement. All rights reserved worldwide.
+ */
 package com.synopsys.integration.common.util.finder;
 
 import java.io.File;
