+++ src/main/java/com/synopsys/integration/detect/tool/detector/executable/DirectoryExecutableFinder.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 
 import org.jetbrains.annotations.Nullable;
 
-import com.synopsys.integration.detectable.detectable.file.FileFinder;
+import com.synopsys.integration.common.util.finder.FileFinder;
 import com.synopsys.integration.util.OperatingSystemType;
 
 // Finds an executable in a directory.
